move model loader logic to regular service
Up until now the loader logic was behind a protected service definition,
which made it convenient to pass in a name of the model to load, but a
pain to extend. This change keeps the protected service as before, but
it sets the name of the model to load to a application property, and
calls the regular loader service which keeps the same logic.