construct order by statement when the select statement is being built
If the order by is built when 'orderBy' is called, the 'table' may not
yet be set. To work around this issue the 'orderBy' now only stores the
data into the internal array, and constructs the 'orderBy' when the
'select' method is called and the select statement is being built, since
the table absolutely has to be set then already.